mondai hakkaku

+++ utils.c
@@ -1,4 +1,4 @@
-#include <stdio.h>
+#include "minishell.h"
 
 void     _bcopy(void *dest, void *src, int size)
 {
@@ -20,7 +20,7 @@ int     _bcopy_int(void *dest, void *src, int size)
 void	_bzero(void	 *s, int n)
 {
 	char *str = (char *)s;
-	size_t i = 0;
+	int i = 0;
 	while (i < n)
 	{
 		str[i] = 0;
